Move model map types to dedicated module
